SHARE-244 Refactoring - Adminarea - Featured Urls
Featured Programs can now be added via URL or ID in the FeaturedProgramAdmin page. Got rid of the redundand FeaturedUrlAdmin.php. Added a checkbox to decide wether to save an URL or a program with its ID. It is also possible to enter the link to a program, when the "Use Url" checkbox is unchecked, the programs ID will be cropped from the link and saved. Also now not crashing anymore when entering a bad id or url.