SHARE-388 Incorrect notification text for remixes
there was an "typo" in the english translation file. instead of '%remix_program_link%' as in all other trans-files there was '%remixed_program_link%'. i changed it in the corresponding JS and TWIG file for notifications and in the english translation file to fit with the other translations.
adjusted test case to test whole notification, not just a part of it