SHARE-407 Project credits/description sync.
When a user uploads a project to the share, the user can add a "Description" and "Notes and credits" before the actual upload. If the owner of the project now updates this information on the share website the updated attributes are only stored in the db and not in the "code.xml" file. "Notes and credits" is a new attribute in the app and is now read out during the upload of the project and stored in the database. Furthermore the code.xml file is now also overwritten when changing the attributes. If a ".catrobat" file already exists for the project, it will be deleted during the update to prevent a caching problem.