SHARE-404 "Help" link under "Useful links" in footer adjusted to wiki link
The help link in the footer as well as in the sidebar was replaced and now points to the wiki where all resources were moved to. All files under /app/help were removed as well as the crowdin keywords which also were removed from the english language file. Old tests were removed or adjusted.