Custom Translation Feature - Editor
Add ability to create a custom translation for the description and credits sections of a user's program through a new language selector in the program editors. When the translation is saved, a snackbar notification is shown to confirm that the translation is saved. Also, add popup to prompt user to keep or discard unsaved changes when langauge selector is changed. Finally, add new popup to prompt users to save their unsaved changes when the program editor is closed.