SHARE-327 Scratch Project Notes and Credits
Implemented displayal on a progam detail page whether it is a Scrath project and if so print out that it is a scratch project and link the original project from scratch.
I changed the twig template for the program description and also updated ProjectDataFixtures in order to be able to test this US appropriately.