SHARE-577 Add "Share" option to project overview context menu
Added a "Share" option to the project overview context menu to make it easier to access the share link. This new button is inserted between the existing "Set public" and "Delete project" options and leads to the same result as clicking the share button in the toolbar on the project page.